Image_generation/requirements.txt
beautifulsoup4
lxml
nltk
fastapi==0.100.0
uvicorn==0.23.0
fastcore
uvicorn
pytest
#llama-cpp-python==0.2.82
pydantic==2.8.2
python-dotenv==1.0.0
torch
diffusers
accelerate
litserve
transformers
streamlit
faiss-cpu==1.12.0
pandas==2.2.3
pillow==11.3.0
nltk==3.9.1
numpy==2.3.2
requests==2.32.4
streamlit==1.48.1
sentence-transformers==5.1.0
transformers==4.55.2